requirements.txt
advertorch==0.2.3
albumentations==1.1.0
colormath==3.0.0
dlib==19.24.0
efficientnet_pytorch==0.7.1
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.4
inplace_abn==1.1.0
kornia==0.6.12
lmdb==1.5.1
loralib==0.1.2
matplotlib==3.5.3
numpy==1.21.5
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==1.3.5
Pillow==9.0.1
psutil==6.1.1
pytorch_msssim==1.0.0
PyYAML==6.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
simplejson==3.19.3
timm==0.6.12
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.65.2
transformers==4.30.2
